docker_python/requirements.txt
apsw<=3.16.2.post1
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.0
chardet==3.0.4
CherryPy==8.9.1
Click==7.0
configparser==3.5.0
cryptography==2.1.4
cssselect==1.0.3
cssutils==1.0.2
defusedxml==0.5.0
dicttoxml==1.7.4
dnspython==1.15.0
enum34==1.1.6
eyeD3==0.8.4
feedparser==5.2.1
Flask==1.0.2
futures==3.2.0
html5-parser==0.4.4
html5lib==0.999999999
idna==2.6
ipaddress==1.0.17
irc>=0.0.8.5.3
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.3.1
lxml==4.2.1
Markdown==2.6.9
MarkupSafe==1.1.0
mccabe==0.6.1
mechanize==0.2.5
msgpack==0.5.6
mutagen==1.38
netifaces==0.10.4
numpy==1.13.3
olefile==0.45.1
pathlib==1.0.1
Pillow==5.1.0
pyAesCrypt==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
Pygments==2.2.0
pyliblzma==0.5.3
pylint==1.9.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
PyPDF2==1.26.0
python-dateutil==2.6.1
python-magic<=0.4.16
pyxdg==0.25
PyYAML==3.12
regex==2017.12.12
repoze.lru==0.7
Routes==2.4.1
scour==0.36
SecretStorage==2.3.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.12.0
SOAPpy==0.12.22
webencodings==0.5
WebOb==1.7.3
Werkzeug==0.14.1
wrapt==1.10.11
wstools==0.4.3
